Moved processing functions to new file
Previously, processing functions were located in Main.hs, which was
getting somewhat cluttered. All of this functionality has now been split
off to a separate file, Processing.hs and includes most of the P2P ()
actions defined in Main.hs, since these are not dependent on the IO ()
and forking stuff handled in Main.hs.