Made protocol and implementation port-independent
Port information is now carried around in the protocl and properly
memorized as well. This allows for a practically unbound number of local
connections being tested at the same time, as required for issue #12.

Port information is carried around in IAM and PEER.

Additionally, a number of semantic changes were performed, such as
making sure PANIC's reply does not send a client's own address to
himself, as this can cause nasty runaway recursion.

Perhaps most importantly, it is now customary to respond with a
collection of PEERs at your disposal when being demanded to IDENTIFY.
Note that this does not represent a final solution, see issue #7 for
more information on this problem. This is just a temporarily
work-around until proper address assignment has been implemented.