Moved homePort to a Reader instead, flattened RWST
Since homePort isn't configurable at runtime it makes sense to put it
into a ReaderT instead of the StateT, along with (eventually) the rest
of the global runtime options, provided we ever implement any other than
the port (which will become a -p flag).

Since, after the addition of a ReaderT, we have a composition of a
ReaderT, WriterT and StateT, I've flattened it to an RWST instead, which
removes some of the unnecessary verbosity - especially from GHC error
messages. This also makes more clear the ordering between ErrorT and
RWST, since the RWST compositions are essentially commutative.