Implemented local message queue for DHT-unknowns
The Queue is simply a list of functions which need either a Name or an
Id to be resolved (to an Id/Address, respectively), after which the
action executes with the required parameter. (The Queue functions
themselves already query the local DHT fragments when passing the
parameters to the lambda)

This partially resolves issue #4, though DHT failures don't drop queue
items yet. Global messages are never used unless explicitly warranted
now.

Note that Id resolves where the address is already known (for whatever
reason, such as DHT mirror overlap) don't trigger the required hasAddr
yet to actually fire the action, this needs to be fixed.