Cleaned up show instance for P2PState
This is mainly to make the output of test.dump considerably cleaner,
thus helping with the debugging process immensely.

Also moved the instance declaration for Main because it doesn't really
make sense anywhere else; not to mention needing the other files as
dependency for eg. id hashing (keys tend to be extremely verbose).

Improvement ideas: Obfuscate public keys more to break the correlation
between the “hash” and actual addresses, since it's only used as a
work-around. That amount of precision isn't needed either.