Implemented DHT lookups and fixed queue delays
DHT requests now get fired off to the appropriate addresses (as given by
the hashing functions in P2P.Math) when a delayed action is put into the
queue. For elements with existing DHT entries, this is now properly
skipped and executed directly.

Cleaned up some of the DHT table interaction functions (eg. getAddr and
getId) as well as some of the functions in P2P.Sending as well, to allow
for a more general form without too much code duplication.

This further implements issue #4, but queue purging is still missing.