Added QUIT section
This section is used to directly signal that a peer will be dropping out
from the network, in order to have other peers forget about them
(similar to the rest of a address routing failure) as well as clearing
the associated AES keys from their caches.

In addition, some of the helper functions in P2P.hs and Sending.hs have
been cleaned up, reordered or relabelled slightly.