Reworked PANIC/PEER to allow for dial-in
This adds the basic framework for remote network reflection, allowing
using an Approx REQUEST to get a handy collection of (Host, Port, Addr)
tuples used for the initial connections. The Address fields are only
given so the bootstrapper can calculate and assign an appropriately
placed address for the joining node.

This partially addresses issue #7 but does not account for near-empty
networks nor does it actually implement the bootstrapping process.
Further food for thought:

Consider transferring an ID field in PEER as well, but note that this
can dangerously allow for third-party editing unless we add signatures
generously. Alternatively, don't bother; since it's not strictly needed.

I removed the odd PEER/IDENTIFY relationship as well, which really makes
little sense in this new context.

Further minor adjustments such as code prettification or shuffling of
helper functions.