Added basic bootstrapping logic
This adds most of the still missing bootstrapping logic, including
adding a queue for the to-be-bootstrapped clients and various other
dial-in logic.

This contributes to issue #7 but (still) does not implement it fully.
Notably still missing is proper Approx routing for empty or near-empty
networks, probably implemented by adding loopback connections when the
network is empty (this will also allow for appropriate recovery when the
network gets reduced to a single peer, instead of just specialcasing the
creation process).