Implemented remaining missing DHT logic
In addition to implementing all of the missing DHT logic, this commits
fixes a vast number of bugs that were not found until now due to testing
issues.

As of now, bootstrapping itself works with at least 3 peers and the
subsequent routing works as expected. This mostly resolves issue #7,
but note that the current implementation is *still* incorrect. It works,
but it's incorrect.

To elaborate, the REQUEST is currently hard-coded to be sent to address
0.7, while it should be sent to a random address. This is considered a
trivial change.

There is another issue lurking in the current implementation: Peers can
be ghosted by a new peer joining the network due to the RESPONSE's sender
not being included in the PEER list. A solution for this is still in
development.