Various bug fixes and changes to bootstrapping
Notable changes:

1) An off by one in the makeUnique calculation code has been resolved.
2) The context is now appropriately cleared when parsing loopbacks.

Still to be done:

After connecting a peer, we should still retrieve addresses directly
from the peer in question. Alternatively, the peer's remote address
should be included in the RESPONSE somehow, but then this needs to be
specialcased for loopbacks. Maybe by terminating or not terminating the
remote connection host-side? That way the missing host information won't
be needed.

Related to issue #7.