Added Channel messaging properly
This amends the protocol slightly to require “AESMSG” to be prefixed in
front of all AES-encrypted messages, to know on the receiving end
whether they were decrypted successfully or not. Also added is a new
list in the P2PState, chanList, against which all incoming Channel
messages will be matched to see if they decode properly.